UserLanguage & UserResponse registration is working!
- created UserLanguageRepository
- created UserResponseRepository
- add more models in the UserRegistrationViewModel
- RegistrationService correctly adds to all User[..] tables

ISSUE:
our Responses dictionary [QuestionID, SelectedResponseOptionID] wasn't being filled by the submission of the form. The Responses came in properly however, there is an issue with the automatically binding which left our Responses dictionary and table empty.
a temporary fix was implemented in RegistrationController.cs (will need to be moved because of design)
Issue could be a problem with ASP.NET..